+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Support `menu_bar`
 - Hello, Menu Bar! Sample
 - Remove the need to call Document.ready? before opening a Glimmer shell
+- Support opening a message_box before creating a shell
 
 ## 0.8.0